Refactor minnisblad routes: move utility functions to utils.py and simplify upload_file logic
